refactor: UpdateChapterRequest 위치 변경 #15
